requirements.txt
transformers~=4.28.1
scikit-learn~=1.4.2
pandas~=2.2.2
pillow~=10.2.0